+++ src/ts/upload/index.ts
@@ -1,6 +1,6 @@
-import {getEditorRange, setSelectionFocus} from "../util/selection";
-import {getElement} from "./getElement";
-import {setHeaders} from "./setHeaders";
+import { getEditorRange, setSelectionFocus } from "../util/selection";
+import { getElement } from "./getElement";
+import { setHeaders } from "./setHeaders";
 
 class Upload {
     public element: HTMLElement;
@@ -16,7 +16,7 @@ class Upload {
 
 const validateFile = (vditor: IVditor, files: File[]) => {
     vditor.tip.hide();
-    const uploadFileList = [];
+    const uploadFileList: File[] = [];
     let errorTip = "";
     let uploadingStr = "";
     const lang: keyof II18n | "" = vditor.options.lang;
@@ -63,11 +63,22 @@ const validateFile = (vditor: IVditor, files: File[]) => {
 
         if (validate) {
             uploadFileList.push(file);
-            uploadingStr += `<li>${filename} ${window.VditorI18n.uploading}</li>`;
+            uploadingStr += `<li>${filename} ${window.VditorI18n.uploading} <a class="vditorCancelUpload" href="javascript:void(0)">${window.VditorI18n.cancelUpload}</a></li>`;
         }
     }
 
     vditor.tip.show(`<ul>${errorTip}${uploadingStr}</ul>`);
+
+    if (vditor.options.upload.cancel) {
+        const vditorCancelUploadElement = vditor.tip.element.querySelector(".vditorCancelUpload");
+        if (vditorCancelUploadElement) {
+            vditorCancelUploadElement.addEventListener("click", () => {
+                vditor.options.upload.cancel(uploadFileList);
+                vditor.tip.hide();
+                vditor.upload.isUploading = false;
+            });
+        }
+    }
 
     return uploadFileList;
 };
@@ -208,6 +219,7 @@ const uploadFiles =
         }
 
         const xhr = new XMLHttpRequest();
+        vditor.upload.xhr = xhr;
         xhr.open("POST", vditor.options.upload.url);
         if (vditor.options.upload.token) {
             xhr.setRequestHeader("X-Upload-Token", vditor.options.upload.token);
@@ -243,6 +255,7 @@ const uploadFiles =
                     element.value = "";
                 }
                 vditor.upload.element.style.display = "none";
+                vditor.upload.xhr = undefined;
             }
         };
         xhr.upload.onprogress = (event: ProgressEvent) => {
@@ -257,4 +270,4 @@ const uploadFiles =
         xhr.send(formData);
     };
 
-export {Upload, uploadFiles};
+export { Upload, uploadFiles };